Moved DragonConfig to its own file
