+++ src/lib/dragon/builder-states/edit-form/FormSectionHPAbilities.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import {
-		type DragonConfig,
 		maxHPMin,
 		maxHPMax,
 		numberOfHitDiceMin,
@@ -8,6 +7,7 @@
 		abilityMin,
 		abilityMax
 	} from '$lib/dragon';
+	import type { DragonConfig } from '$lib/dragon/dragon-config';
 
 	export let config: DragonConfig;
 </script>
