+++ src/lib/dragon/builder-states/edit-form/FormSectionSkills.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import { type DragonConfig, SKILLS } from '$lib/dragon';
+	import { SKILLS } from '$lib/dragon';
+	import type { DragonConfig } from '$lib/dragon/dragon-config';
 
 	export let config: DragonConfig;
 </script>
