+++ src/lib/dragon/builder-states/edit-form/FormSubsectionWelcome.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import { type DragonConfig, COLORS, COLORS_UPPER, AGES, AGES_UPPER } from '$lib/dragon';
+	import { COLORS, COLORS_UPPER, AGES, AGES_UPPER } from '$lib/dragon';
+	import type { DragonConfig } from '$lib/dragon/dragon-config';
 
 	export let config: DragonConfig;
 </script>
